Removed distinct scope from many model associations
The distinct scope is only really useful for model associations
(has_many etc) defined with a through: argument, so that a single entity
appears only one in the collection even if there are multiple "paths" to
it through several entities.

So, the removal of the distinct scope for associations between models
which are not defined with a through: clause should not change behavior.
SQLs will be executed without an (unnecessary) DISTINCT clause, which
should even be a bit faster.

Also entities are guaranteed not to have duplicates in the tables
because of the validations set in the models.

For more about the distinct scope and why it was not doing anything here
see:

http://guides.rubyonrails.org/association_basics.html#has-many-association-reference