Faster model deletion in some cases.
Instead of dependent: :destroy, in some model associations dependent:
:delete_all is used. This makes deletions faster because associated
entities are deleted in a single SQL statement, instead of instantiating
a rails model and then invoking #destroy on it.

Care must be taken: only models that have no dependent objects or
callbacks that are invoked when destroyed can be deleted this way. In
most cases :destroy must be used to ensure that callbacks are invoked
and dependent objects destroyed. If in the future any of the models that
are currently deleted with :destroy_all have such a callback or
dependent collection added, the change should be reverted and :destroy
used instead in the parent model.