Include only aws-sdk-s3 gem in bundle
Since aws-sdk v3 the aws-sdk-core gem has been broken into multiple gems for the different AWS services. The suggested migration path is to replace aws-sdk-core (or aws-sdk-resources, which is what feedbunch was using) with just the specific gems for the AWS services actually used. In our case we only use S3.