Deal with feed URLs that begin with a mismatched " character
This should fix an error in production; this happens because of a misconfigured feed, but it's easy to fix in our side so feedbunch can use the feed.

Also if an error is raised when attempting to normalize a feed url, rescue the error and replace the problematic url with a blank string.