Reverted 865969df53d5c54b2443769772f6f709e7a7899b
Using 1/0 instead of 't'/'f' during testing causes errors. It seems when
inserting new EntryState records from the Entry#set_unread_state
callback (called in the after_create of Entry), the field state of the
new EntryState instance is set to false but a 'f' is actually inserted
in the test DB (sqlite), regardless of the config option that should
make ActiveRecord insert a 0 there. Afterwards when querying for
EntryStates, the SQL query is built like "where read=0", which finds no
records and makes the test fail.

Something similar may happen with other model classes.

I'm not sure if I'm missing a config option during testing to treat
booleans as 1/0 during testing, or if it's a bug in Rails. For now I
have to live with the deprecation warning during testing.