Monkeypatch to disable advisory locks in postgres
Since this commit:

https://github.com/rails/rails/pull/22122

rails tries to acquire an advisory lock when running migrations; this is
intended to avoid the (unlikely) case of multiple migrations running
concurrently.

However this fails when the app is accessing a Postgres db through
pgbounce in transaction pooling mode (which is the only mode that gives
reasonable memory gains while allowing transactions to work correctly).
Unfortunately FeedBunch is impacted by this. This means that capistrano
deployments fail when trying to run migrations, even if no migration is
actually part of the deployment.

The current ugly fix is a monkey patch to disable advisory lock support
in the Postgres adapter. This is messy and could break other things,
I'm not sure if advisory locks are used elsewhere in Rails.

A discussion is currently underway about adding a config option to
disable the acquisiton of an advisory lock before running a migration,
see:

https://github.com/rails/rails/issues/31190

If such an option is implemented this patch should be removed at once.