WIP: config option to fetch some special feeds differently
This is intended to fix problems when fetching Tumblr feeds; since
recently Tumblr sends an interstitial page about GDPR compliance instead
of the actual feeds, which breaks feed clients like FeedBunch.

I'm going to extend the already existing "special feeds" functionality
to cover this case. The "special feeds" functionality, which was
implemented for Demonoid but can be used for other feeds, fetches feeds
normally but makes some changes to entries just before saving (in the
Demonoid case, to avoid duplicate entries). This is going to be expanded
so that each feed can have both a special fetcher (which fetches a feed
in some different way than a dumb HTTP client) and a special handler
(which makes changes to entries just before saving them).