Implemented new method SpecialFeedManager.get_special_fetcher
This method returns the special fetcher class for a feed configured in
config/special_feeds.yml if any, nil otherwise.

The private code in SpecialFeedManager is as DRY as possible, using both
the url and fetch_url for each feed, and looking into the
special_feeds_fetchers and special_feeds_handlers list in Rails config.