When fetching a feed, use a special fetcher if configured
FeedClient has been reorganized to move the code that performs the
default fetching to the new default_fetch method. This method attempts
to fetch the feed with the simple HTTP client (RestClient) and if this
fails and it seems like the URL is behind Cloudflare DDoS protection,
makes a second attempt with a headless chrome in the hope that a full
featured browser can bypass the DDoS protection.