Use a special class to fetch Tumblr feeds
This class gets around the interstitial GDPR page served by Tumblr by
using a full-featured headless browser to fetch the interstitial page,
click on the OK button and wait for the actual content (either webpage
or feed XML) to actually load.

The mechanism is extensible to other classes that need special fetchers
in the future.