Better autoscrolling when opening entries at the bottom of the list.
The autoscroll now detects if the open entry is not at the top of the
viewport. This happens often with the entries at the bottom of the
entries list. When this happens, the height of an empty div at the
bottom is increased to push the entry upwards, so that it is displayed
at the top of the viewport.

This way when opening an entry the autoscroll always positions it at the
same position (the top of the viewport, just under the navbar)
regardless of the position of the entry in the list.