Determine if an entry needs a special handler based on its feed url.
Previously the feed's fetch_url field was also used to see if there was
a match in the special feeds list. This is not a good idea: websites can
change where they host their feeds without notice. Looking at the
website url should be enough to know if it's a special website with a
feed that needs some massaging to be accepted by feedbunch, the URL at
which the feed itself is hosted is immaterial.