Use Firefox instead of Chrome to bypass Cloudflare DDOS protection.
When a site is behind Cloudflare DDOS protection, FeedBunch uses a
full headless browser to bypass it because a simple http client like
RestClient is stopped by the challenge page.

I was using headless Chrome for this. It worked fine until recently,
when Cloudflare started treating headless Chrome differently for some
sites. Instead of issuing a javascript challenge, which headless Chrome
can bypass, Cloudflare issues a reCaptcha challenge with cannot be
bypassed without human intervention. This means RSS feeds for those
sites stopped working. I realized that a fully automated (but not
headless) Chrome browser is issued a js challenge instead of a
reCaptcha, which means that Cloudflare somehow detects we're using a
headless Chrome and making it impossible for us to bypass the challenge
page, even though FeedBunch is attempting to make a legitimate use of
the Cloudflare-protected site.

The fix is to use Firefox instead of Chrome. For now Cloudflare treats
full Firefox and headless Firefox the same, issuing a js challenge that
FeedBunch can bypass, making those RSS feeds operative again.

Note.- For headless Firefox to work, the "geckodriver" binary must be in
the server PATH. For some Linux distros it cannot be installed from the
package manager but it must be downloaded from github and copied to a
folder in the PATH instead. This means geckodriver won't get updated
automatically by the package manager. Not happy with this but I don't
see a solution for now.