Detect duplicated entries based on content, not just guid
Unfortunately some feed owners are less than careful and change the
way guids are generated for already published entries. This means that
sometimes FeedBunch fetches an entry (or, usually, many in the same feed)
which was already previously fetched, but with a different guid. This
results in FeedBunch identifying it as a new entry (because the guid
does not match any guid already saved in the db for that feed) and
therefore saving it as a new entry in the database. Of course this new
entry will default to unread state for all subscribed users.

From the end-user point of view this results in duplicated entries, or
already read entries "reappearing" in the unread entries list.

This is totally the feed owner's fault, and the way FeedBunch behaves is
completely according to spec. But this is little consolation for the
users, who have to deal with unwanted duplications.

The workaround is to identify entry uniqueness not just based on entry
guid, but also on the content of each entry. When fetching feeds an
entry will be treated as a duplicate and ignored if either its guid or
its content matches that of an entry of the same feed already saved in
the database.

For performance reasons FeedBunch does not compare the full entry
contents. Instead an MD5 hash of the content is calculated and used to
detect duplicates. Each entry now has its content hash saved in the
database. MD5 is good enough for this purpose because we're not using it
for securing anything, just to detect duplicated content. There's the
risk of false positives because of hash collisions, but with a low
likelihood.

Not yet implemented: taking old deleted entries into account when
detecting duplicates based on content hash.