More accurate entry hashing
The hash that uniquely identifies an entry in the scope of a feed is now
calculated using the concatenation of its content (if present), summary
(if present) and title (which is always present).

This helps reduce the number of false positives in feeds that often
repeat the same content in entries; e.g. torrent tracker sites which
often have just the magnet link in the entry url, with no content nor
summary.