Use id to order entries in migration
When deleting entries with duplicate hashes in this migration it is
important to leave the oldest entry alone and delete any newer
duplicates. The id attribute is used in the case of duplicate entries
with the same publish date and created_at date; in this case the entry
with the lowest id will be left alone and the rest will be considered
duplicates and deleted.