Fixed migration
Now it doesn't invoke activerecord validations or callbacks, which can
fail if the model class is out of sync with the db schema when running
the migration. Instead we use methods that update the db directly.

Also entries are processed in batches of 1000, instead of all at once,
which can easily run into memory shortage problems.