Rewritten migrations to update db indexes
The addition of the unique_hash column to entries and deleted_entries
tables means it's a good idea to drop a couple old db indexes and create
new ones adding the unique_hash column in each one.

The migration that sets md5 hashes for all old entries should also run
much faster in production after this.