Faster deletion of users and feeds
When deleting a user or a feed, many depending objects are mass-deleted
with a single SQL sentence (bypassing activerecord callbacks and
validations), instead of instantiating each model and deleting each
object one by one. This should improve speed when deleting a user or
when a user unsubscribes from a feed and there are no more subscribed
users, which triggers feed deletion.

Also some of the logic that is triggered when deleting a
feed_subscription has been moved from callbacks in the User model to
callbacks in the FeedSubscription model. It's a more logical place and
should make the code easier to maintain.