Associate validation errors to base object, not a specific attribute
When trying to save a new deleted_entries instance a validation checks
that there isn't a non-deleted entry with the same feed_id and either
guid of unique_hash. Similarly when trying to save a new entry instance
a validation checks that there isn't a deleted entry with the same
feed_id and either guid or unique_hash.

In both cases if the validation fails, the error is not attributable to
a single model attribute and must be associated with the base object
instead.