Earlier detection of entries with duplicate content during fetching
Until now the detection of entries with duplicated content (based on the
md5 hash of the concatenation of content, summary and title) happened in
an ActiveRecord validation just before saving each entry. This meant
that when a duplicated entry was detected, it was not saved in the
database (as desired) but an error was raised and logged.

Raising an error in this case is too harsh. Too many validation errors
appear in the application log. If an entry has changed the way guids are
generated (e.g. if the entry URL is part of the guid and the site owner
has changed from http to https) probably we already have many entries
with the old guis in the db, but every time we fetch the feed those
entries will come with the new guids. This means that they would only be
detected as duplicates in the validation and a lot of validation errors
would be written to the log periodically, every time the feed is
fetched.

As a general rule, I don't like having errors written to the log
periodically when nothing's actually failing.

The feed fetching and entry validation logic has been rewritten a bit.
Now the EntryManager class is in charge of invoking methods on the
just-created but not-yet-saved Entry instance, to check if there are
other entries (or deleted entries) for the same feed and with either the
same guid or the same unique_hash. If a duplicated is detected at this
poing a message is written to the log with debug priority and the next
entry is processed; otherwise the entry is saved to the db.

EntryManager was already detecting duplicates based on the guid. Now it
also detects duplicates based on unique_hash. It's a logical extension
of its responsabilities. Some rewriting of the Entry model has been
necessary because, unlike the gui, we don't have the entry unique_hash
in the parsed XML, we have to actually build an Entry model instance
(without saving it) to calculate its unique_hash.