Enable following redirects in RestClient.
RestClient should follow redirects automatically, but I'm getting errors
when a server returns a 301 code. It's probably caused by the
rest-client-components gem, which I need for client-side HTTP caching;
it's mostly unmaintained and it seems it breaks the redirect-following
feature of RestClient somewhat.

Explicitly adding a parameter to RestClient.get to indicate the maximum
number of redirects to follow before throwing an error seems to fix it.