Mark belongs_to association as optional
The belongs_to relationship from the SubscribeJobState model to the Feed
model is optional; only if the job state is SUCCESS does the object
actually belong to a feed, otherwise SubscribeJobState#feed_id is nil.

Since 25d4a002329cf4d1173a62af16d022b89e0fadde FeedBunch loads the Rails
5.0 config defaults in config/application.rb and among those faults is a
flag that raises an error if a model is saved with a nil at the other
end of a belongs_to association, unless the :optional flag is set to
true in the association.