The FORCE_SSL env variable can make the app secure.
The app is secure by default, if FORCE_SSL is not set.

This toggles the force_ssl rails flag, which among other things makes
TLS mandatory, marks cookies as secure-only and enables HSTS. Generally
this should be enabled in production, but it's useful to disable it in
some use cases.