Use a single sidekiq server initialization block
FeedBunch was making two calls to Sidekiq.configure_server : one to
initialize the redis connection pool and a second one to create default
scheduled jobs. There is a possibility that this was causing a bug in
which Sidekiq always tried to connect to Redis in localhost:6379,
regardless of what was in the configuration. Hopefully using a single
configure_server call avoids this.