FeedBunch can use a remote browser instead of a local headless browser
When fetching certain feeds a full browser is necessary (e.g. to pass
certain anti-DDoS measures). Until now a local Chromium was started
every time such a feed was fetched, with required chromium and its
webdriver to be installed in the local machine. This can be problematic
when running FeedBunch in a dockerized environment, because running a
chromium browser process in a container is not trivial.

With these changes a remote browser can be used instead. The HEADLESS_BROWSER_LOCATION
environment variable controls whether a local (value "local") or remote
(value "remote") browser is used. The default is to use a local browser.

If a remote browser is used, HEADLESS_BROWSER_HOST env variable controls
its host (defaults to 127.0.0.1) and the HEADLESS_BROWSER_PORT env
variable controls the port where it's listening for Selenium connections
(defaults to 4444).

The easiest way to get a standalone chrome browser that listens for
selenium connections is to run the selenium/standalone-chrome docker
image.